merge

+++ Logic/source/AI/TriggerManager.cpp
@@ -1,4 +1,5 @@
 #include <AI/TriggerManager.h>
+#include <Physics\Physics.h>
 using namespace Logic;
 
 TriggerManager::TriggerManager() 
@@ -48,7 +49,7 @@ void TriggerManager::update(float deltaTime)
 		
 		// Remove triggers
 		if (t->getShouldRemove())
-			removeTrigger(t, i);
+			removeTrigger(t, (int)i);
 	}
 }
